+++ generators/graph/voxel_generator_graph.cpp
@@ -25,8 +25,9 @@ const char *VoxelGeneratorGraph::SIGNAL_NODE_NAME_CHANGED = "node_name_changed";
 
 VoxelGeneratorGraph::VoxelGeneratorGraph() {
 	_main_function.instantiate();
-	_main_function->connect(VoxelStringNames::get_singleton().changed,
-			callable_mp(this, &VoxelGeneratorGraph::_on_subresource_changed));
+	_main_function->connect(
+			VoxelStringNames::get_singleton().changed, callable_mp(this, &VoxelGeneratorGraph::_on_subresource_changed)
+	);
 }
 
 VoxelGeneratorGraph::~VoxelGeneratorGraph() {
@@ -127,9 +128,15 @@ bool VoxelGeneratorGraph::is_using_xz_caching() const {
 // TODO Optimization: generating indices and weights on every voxel of a block might be avoidable
 // Instead, we could only generate them near zero-crossings, because this is where materials will be seen.
 // The problem is that it's harder to manage at the moment, to support edited blocks and LOD...
-void VoxelGeneratorGraph::gather_indices_and_weights(Span<const WeightOutput> weight_outputs,
-		const pg::Runtime::State &state, Vector3i rmin, Vector3i rmax, int ry, VoxelBuffer &out_voxel_buffer,
-		FixedArray<uint8_t, 4> spare_indices) {
+void VoxelGeneratorGraph::gather_indices_and_weights(
+		Span<const WeightOutput> weight_outputs,
+		const pg::Runtime::State &state,
+		Vector3i rmin,
+		Vector3i rmax,
+		int ry,
+		VoxelBuffer &out_voxel_buffer,
+		FixedArray<uint8_t, 4> spare_indices
+) {
 	ZN_PROFILE_SCOPE();
 
 	// TODO Optimization: exclude up-front outputs that are known to be zero?
@@ -264,8 +271,14 @@ constexpr inline uint16_t make_encoded_indices_for_single_texture(uint8_t index)
 
 // TODO Optimization: this is a simplified output using a complex system.
 // We should implement a texturing system that knows each voxel has a single texture.
-void gather_indices_and_weights_from_single_texture(unsigned int output_buffer_index, const pg::Runtime::State &state,
-		Vector3i rmin, Vector3i rmax, int ry, VoxelBuffer &out_voxel_buffer) {
+void gather_indices_and_weights_from_single_texture(
+		unsigned int output_buffer_index,
+		const pg::Runtime::State &state,
+		Vector3i rmin,
+		Vector3i rmax,
+		int ry,
+		VoxelBuffer &out_voxel_buffer
+) {
 	ZN_PROFILE_SCOPE();
 
 	const pg::Runtime::Buffer &buffer = state.get_buffer(output_buffer_index);
@@ -287,8 +300,17 @@ void gather_indices_and_weights_from_single_texture(unsigned int output_buffer_i
 }
 
 template <typename F, typename Data_T>
-void fill_zx_sdf_slice(Span<Data_T> channel_data, float sdf_scale, Vector3i rmin, Vector3i rmax, int ry, int x_stride,
-		const float *src_data, Vector3i buffer_size, F convert_func) {
+void fill_zx_sdf_slice(
+		Span<Data_T> channel_data,
+		float sdf_scale,
+		Vector3i rmin,
+		Vector3i rmax,
+		int ry,
+		int x_stride,
+		const float *src_data,
+		Vector3i buffer_size,
+		F convert_func
+) {
 	unsigned int src_i = 0;
 	for (int rz = rmin.z; rz < rmax.z; ++rz) {
 		unsigned int dst_i = Vector3iUtil::get_zxy_index(Vector3i(rmin.x, ry, rz), buffer_size);
@@ -300,15 +322,23 @@ void fill_zx_sdf_slice(Span<Data_T> channel_data, float sdf_scale, Vector3i rmin
 	}
 }
 
-void fill_zx_sdf_slice(const pg::Runtime::Buffer &sdf_buffer, VoxelBuffer &out_buffer, unsigned int channel,
-		VoxelBuffer::Depth channel_depth, float sdf_scale, Vector3i rmin, Vector3i rmax, int ry) {
+void fill_zx_sdf_slice(
+		const pg::Runtime::Buffer &sdf_buffer,
+		VoxelBuffer &out_buffer,
+		unsigned int channel,
+		VoxelBuffer::Depth channel_depth,
+		float sdf_scale,
+		Vector3i rmin,
+		Vector3i rmax,
+		int ry
+) {
 	ZN_PROFILE_SCOPE_NAMED("Copy SDF to block");
 
 	if (out_buffer.get_channel_compression(channel) != VoxelBuffer::COMPRESSION_NONE) {
 		out_buffer.decompress_channel(channel);
 	}
 	Span<uint8_t> channel_bytes;
-	ERR_FAIL_COND(!out_buffer.get_channel_raw(channel, channel_bytes));
+	ERR_FAIL_COND(!out_buffer.get_channel_as_bytes(channel, channel_bytes));
 	const Vector3i buffer_size = out_buffer.get_size();
 	const unsigned int x_stride = Vector3iUtil::get_zxy_index(Vector3i(1, 0, 0), buffer_size) -
 			Vector3iUtil::get_zxy_index(Vector3i(0, 0, 0), buffer_size);
@@ -316,22 +346,50 @@ void fill_zx_sdf_slice(const pg::Runtime::Buffer &sdf_buffer, VoxelBuffer &out_b
 	switch (channel_depth) {
 		case VoxelBuffer::DEPTH_8_BIT:
 			fill_zx_sdf_slice(
-					channel_bytes, sdf_scale, rmin, rmax, ry, x_stride, sdf_buffer.data, buffer_size, snorm_to_s8);
+					channel_bytes, sdf_scale, rmin, rmax, ry, x_stride, sdf_buffer.data, buffer_size, snorm_to_s8
+			);
 			break;
 
 		case VoxelBuffer::DEPTH_16_BIT:
-			fill_zx_sdf_slice(channel_bytes.reinterpret_cast_to<uint16_t>(), sdf_scale, rmin, rmax, ry, x_stride,
-					sdf_buffer.data, buffer_size, snorm_to_s16);
+			fill_zx_sdf_slice(
+					channel_bytes.reinterpret_cast_to<uint16_t>(),
+					sdf_scale,
+					rmin,
+					rmax,
+					ry,
+					x_stride,
+					sdf_buffer.data,
+					buffer_size,
+					snorm_to_s16
+			);
 			break;
 
 		case VoxelBuffer::DEPTH_32_BIT:
-			fill_zx_sdf_slice(channel_bytes.reinterpret_cast_to<float>(), sdf_scale, rmin, rmax, ry, x_stride,
-					sdf_buffer.data, buffer_size, [](float v) { return v; });
+			fill_zx_sdf_slice(
+					channel_bytes.reinterpret_cast_to<float>(),
+					sdf_scale,
+					rmin,
+					rmax,
+					ry,
+					x_stride,
+					sdf_buffer.data,
+					buffer_size,
+					[](float v) { return v; }
+			);
 			break;
 
 		case VoxelBuffer::DEPTH_64_BIT:
-			fill_zx_sdf_slice(channel_bytes.reinterpret_cast_to<double>(), sdf_scale, rmin, rmax, ry, x_stride,
-					sdf_buffer.data, buffer_size, [](double v) { return v; });
+			fill_zx_sdf_slice(
+					channel_bytes.reinterpret_cast_to<double>(),
+					sdf_scale,
+					rmin,
+					rmax,
+					ry,
+					x_stride,
+					sdf_buffer.data,
+					buffer_size,
+					[](double v) { return v; }
+			);
 			break;
 
 		default:
@@ -341,8 +399,15 @@ void fill_zx_sdf_slice(const pg::Runtime::Buffer &sdf_buffer, VoxelBuffer &out_b
 }
 
 template <typename F, typename Data_T>
-void fill_zx_integer_slice(Span<Data_T> channel_data, Vector3i rmin, Vector3i rmax, int ry, int x_stride,
-		const float *src_data, Vector3i buffer_size) {
+void fill_zx_integer_slice(
+		Span<Data_T> channel_data,
+		Vector3i rmin,
+		Vector3i rmax,
+		int ry,
+		int x_stride,
+		const float *src_data,
+		Vector3i buffer_size
+) {
 	unsigned int src_i = 0;
 	for (int rz = rmin.z; rz < rmax.z; ++rz) {
 		unsigned int dst_i = Vector3iUtil::get_zxy_index(Vector3i(rmin.x, ry, rz), buffer_size);
@@ -354,15 +419,22 @@ void fill_zx_integer_slice(Span<Data_T> channel_data, Vector3i rmin, Vector3i rm
 	}
 }
 
-void fill_zx_integer_slice(const pg::Runtime::Buffer &src_buffer, VoxelBuffer &out_buffer, unsigned int channel,
-		VoxelBuffer::Depth channel_depth, Vector3i rmin, Vector3i rmax, int ry) {
+void fill_zx_integer_slice(
+		const pg::Runtime::Buffer &src_buffer,
+		VoxelBuffer &out_buffer,
+		unsigned int channel,
+		VoxelBuffer::Depth channel_depth,
+		Vector3i rmin,
+		Vector3i rmax,
+		int ry
+) {
 	ZN_PROFILE_SCOPE_NAMED("Copy integer data to block");
 
 	if (out_buffer.get_channel_compression(channel) != VoxelBuffer::COMPRESSION_NONE) {
 		out_buffer.decompress_channel(channel);
 	}
 	Span<uint8_t> channel_bytes;
-	ERR_FAIL_COND(!out_buffer.get_channel_raw(channel, channel_bytes));
+	ERR_FAIL_COND(!out_buffer.get_channel_as_bytes(channel, channel_bytes));
 	const Vector3i buffer_size = out_buffer.get_size();
 	const unsigned int x_stride = Vector3iUtil::get_zxy_index(Vector3i(1, 0, 0), buffer_size) -
 			Vector3iUtil::get_zxy_index(Vector3i(0, 0, 0), buffer_size);
@@ -373,18 +445,39 @@ void fill_zx_integer_slice(const pg::Runtime::Buffer &src_buffer, VoxelBuffer &o
 			break;
 
 		case VoxelBuffer::DEPTH_16_BIT:
-			fill_zx_integer_slice<uint16_t>(channel_bytes.reinterpret_cast_to<uint16_t>(), rmin, rmax, ry, x_stride,
-					src_buffer.data, buffer_size);
+			fill_zx_integer_slice<uint16_t>(
+					channel_bytes.reinterpret_cast_to<uint16_t>(),
+					rmin,
+					rmax,
+					ry,
+					x_stride,
+					src_buffer.data,
+					buffer_size
+			);
 			break;
 
 		case VoxelBuffer::DEPTH_32_BIT:
-			fill_zx_integer_slice<uint32_t>(channel_bytes.reinterpret_cast_to<uint32_t>(), rmin, rmax, ry, x_stride,
-					src_buffer.data, buffer_size);
+			fill_zx_integer_slice<uint32_t>(
+					channel_bytes.reinterpret_cast_to<uint32_t>(),
+					rmin,
+					rmax,
+					ry,
+					x_stride,
+					src_buffer.data,
+					buffer_size
+			);
 			break;
 
 		case VoxelBuffer::DEPTH_64_BIT:
-			fill_zx_integer_slice<uint64_t>(channel_bytes.reinterpret_cast_to<uint64_t>(), rmin, rmax, ry, x_stride,
-					src_buffer.data, buffer_size);
+			fill_zx_integer_slice<uint64_t>(
+					channel_bytes.reinterpret_cast_to<uint64_t>(),
+					rmin,
+					rmax,
+					ry,
+					x_stride,
+					src_buffer.data,
+					buffer_size
+			);
 			break;
 
 		default:
@@ -497,8 +590,13 @@ VoxelGenerator::Result VoxelGeneratorGraph::generate_block(VoxelGenerator::Voxel
 
 				// Do a quick analysis of the area. We'll only compute voxels if necessary.
 				{
-					QueryInputs<math::Interval> range_inputs(*runtime_ptr, math::Interval(gmin.x, gmax.x),
-							math::Interval(gmin.y, gmax.y), math::Interval(gmin.z, gmax.z), sdf_input_range);
+					QueryInputs<math::Interval> range_inputs(
+							*runtime_ptr,
+							math::Interval(gmin.x, gmax.x),
+							math::Interval(gmin.y, gmax.y),
+							math::Interval(gmin.z, gmax.z),
+							sdf_input_range
+					);
 					runtime.analyze_range(cache.state, range_inputs.get());
 				}
 
@@ -586,7 +684,8 @@ VoxelGenerator::Result VoxelGeneratorGraph::generate_block(VoxelGenerator::Voxel
 
 				if (_use_optimized_execution_map) {
 					runtime.generate_optimized_execution_map(
-							cache.state, cache.optimized_execution_map, to_span(required_outputs), false);
+							cache.state, cache.optimized_execution_map, to_span(required_outputs), false
+					);
 				}
 
 				{
@@ -621,38 +720,51 @@ VoxelGenerator::Result VoxelGeneratorGraph::generate_block(VoxelGenerator::Voxel
 					// Full query (unless using execution map)
 					{
 						QueryInputs query_inputs(*runtime_ptr, x_cache, y_cache, z_cache, input_sdf_slice_cache);
-						runtime.generate_set(cache.state, query_inputs.get(), _use_xz_caching && ry != rmin.y,
-								_use_optimized_execution_map ? &cache.optimized_execution_map : nullptr);
+						runtime.generate_set(
+								cache.state,
+								query_inputs.get(),
+								_use_xz_caching && ry != rmin.y,
+								_use_optimized_execution_map ? &cache.optimized_execution_map : nullptr
+						);
 					}
 
 					if (sdf_output_buffer_index != -1
-							// If SDF was found uniform, we already filled the results, and we did not require it in the
-							// query. But if another output exists, a query might still run (so we end up at this
-							// `if`), and we should not gather SDF results. Otherwise it would overwrite the slice with
-							// garbage since SDF was skipped.
-							// The same logic goes for other outputs: if they aren't in the query, we must not fill
-							// them.
-							&& !sdf_is_uniform) {
+						// If SDF was found uniform, we already filled the results, and we did not require it in the
+						// query. But if another output exists, a query might still run (so we end up at this
+						// `if`), and we should not gather SDF results. Otherwise it would overwrite the slice with
+						// garbage since SDF was skipped.
+						// The same logic goes for other outputs: if they aren't in the query, we must not fill
+						// them.
+						&& !sdf_is_uniform) {
 						const pg::Runtime::Buffer &sdf_buffer = cache.state.get_buffer(sdf_output_buffer_index);
 						fill_zx_sdf_slice(
-								sdf_buffer, out_buffer, sdf_channel, sdf_channel_depth, sdf_scale, rmin, rmax, ry);
+								sdf_buffer, out_buffer, sdf_channel, sdf_channel_depth, sdf_scale, rmin, rmax, ry
+						);
 					}
 
 					if (type_output_buffer_index != -1 && !type_is_uniform) {
 						const pg::Runtime::Buffer &type_buffer = cache.state.get_buffer(type_output_buffer_index);
 						fill_zx_integer_slice(
-								type_buffer, out_buffer, type_channel, type_channel_depth, rmin, rmax, ry);
+								type_buffer, out_buffer, type_channel, type_channel_depth, rmin, rmax, ry
+						);
 					}
 
 					if (runtime_ptr->single_texture_output_index != -1 && !single_texture_is_uniform) {
-						gather_indices_and_weights_from_single_texture(runtime_ptr->single_texture_output_buffer_index,
-								cache.state, rmin, rmax, ry, out_buffer);
+						gather_indices_and_weights_from_single_texture(
+								runtime_ptr->single_texture_output_buffer_index, cache.state, rmin, rmax, ry, out_buffer
+						);
 					}
 
 					if (runtime_ptr->weight_outputs_count > 0) {
 						gather_indices_and_weights(
 								to_span_const(runtime_ptr->weight_outputs, runtime_ptr->weight_outputs_count),
-								cache.state, rmin, rmax, ry, out_buffer, spare_texture_indices);
+								cache.state,
+								rmin,
+								rmax,
+								ry,
+								out_buffer,
+								spare_texture_indices
+						);
 					}
 				}
 			}
@@ -731,8 +843,13 @@ bool VoxelGeneratorGraph::generate_broad_block(VoxelGenerator::VoxelQueryData &i
 	const Vector3i gmax = origin + (bs << input.lod);
 
 	{
-		QueryInputs<math::Interval> range_inputs(*runtime_ptr, math::Interval(gmin.x, gmax.x),
-				math::Interval(gmin.y, gmax.y), math::Interval(gmin.z, gmax.z), math::Interval());
+		QueryInputs<math::Interval> range_inputs(
+				*runtime_ptr,
+				math::Interval(gmin.x, gmax.x),
+				math::Interval(gmin.y, gmax.y),
+				math::Interval(gmin.z, gmax.z),
+				math::Interval()
+		);
 		runtime.analyze_range(cache.state, range_inputs.get());
 	}
 
@@ -800,7 +917,10 @@ bool VoxelGeneratorGraph::generate_broad_block(VoxelGenerator::VoxelQueryData &i
 namespace {
 
 bool has_output_type(
-		const pg::Runtime &runtime, const ProgramGraph &graph, pg::VoxelGraphFunction::NodeTypeID node_type_id) {
+		const pg::Runtime &runtime,
+		const ProgramGraph &graph,
+		pg::VoxelGraphFunction::NodeTypeID node_type_id
+) {
 	for (unsigned int other_output_index = 0; other_output_index < runtime.get_output_count(); ++other_output_index) {
 		const pg::Runtime::OutputInfo output = runtime.get_output_info(other_output_index);
 		const ProgramGraph::Node &node = graph.get_node(output.node_id);
@@ -1064,9 +1184,15 @@ void VoxelGeneratorGraph::generate_series(Span<float> in_x, Span<float> in_y, Sp
 	// matters if we are storing it inside 16-bit or 8-bit VoxelBuffer.
 }
 
-void VoxelGeneratorGraph::generate_series(Span<const float> positions_x, Span<const float> positions_y,
-		Span<const float> positions_z, unsigned int channel, Span<float> out_values, Vector3f min_pos,
-		Vector3f max_pos) {
+void VoxelGeneratorGraph::generate_series(
+		Span<const float> positions_x,
+		Span<const float> positions_y,
+		Span<const float> positions_z,
+		unsigned int channel,
+		Span<float> out_values,
+		Vector3f min_pos,
+		Vector3f max_pos
+) {
 	std::shared_ptr<Runtime> runtime_ptr;
 	{
 		RWLockRead rlock(_runtime_lock);
@@ -1102,8 +1228,11 @@ void VoxelGeneratorGraph::generate_series(Span<const float> positions_x, Span<co
 		float *ptr_x = const_cast<float *>(positions_x.data());
 		float *ptr_y = const_cast<float *>(positions_y.data());
 		float *ptr_z = const_cast<float *>(positions_z.data());
-		generate_set(Span<float>(ptr_x, positions_x.size()), Span<float>(ptr_y, positions_y.size()),
-				Span<float>(ptr_z, positions_z.size()));
+		generate_set(
+				Span<float>(ptr_x, positions_x.size()),
+				Span<float>(ptr_y, positions_y.size()),
+				Span<float>(ptr_z, positions_z.size())
+		);
 	}
 
 	const pg::Runtime::Buffer &buffer = get_tls_cache().state.get_buffer(buffer_index);
@@ -1192,8 +1321,14 @@ void VoxelGeneratorGraph::bake_sphere_bumpmap(Ref<Image> im, float ref_radius, f
 		const float sdf_min;
 		const float sdf_max;
 
-		ProcessChunk(pg::Runtime::State &p_state, const Runtime &p_runtime, float p_ref_radius, float p_sdf_min,
-				float p_sdf_max, Image &p_im) :
+		ProcessChunk(
+				pg::Runtime::State &p_state,
+				const Runtime &p_runtime,
+				float p_ref_radius,
+				float p_sdf_min,
+				float p_sdf_max,
+				Image &p_im
+		) :
 				im(p_im),
 				runtime_wrapper(p_runtime),
 				state(p_state),
@@ -1295,8 +1430,13 @@ void VoxelGeneratorGraph::bake_sphere_normalmap(Ref<Image> im, float ref_radius,
 		const float strength;
 		const float ref_radius;
 
-		ProcessChunk(pg::Runtime::State &p_state, Image &p_im, const Runtime &p_runtime_wrapper, float p_strength,
-				float p_ref_radius) :
+		ProcessChunk(
+				pg::Runtime::State &p_state,
+				Image &p_im,
+				const Runtime &p_runtime_wrapper,
+				float p_strength,
+				float p_ref_radius
+		) :
 				im(p_im),
 				runtime_wrapper(p_runtime_wrapper),
 				state(p_state),
@@ -1435,8 +1575,14 @@ bool VoxelGeneratorGraph::get_shader_source(ShaderSourceData &out_data) const {
 	StdString code_utf8;
 	StdVector<pg::ShaderParameter> params;
 	StdVector<pg::ShaderOutput> outputs;
-	pg::CompilationResult result = pg::generate_shader(graph, _main_function->get_input_definitions(), code_utf8,
-			params, outputs, Span<const pg::VoxelGraphFunction::NodeTypeID>());
+	pg::CompilationResult result = pg::generate_shader(
+			graph,
+			_main_function->get_input_definitions(),
+			code_utf8,
+			params,
+			outputs,
+			Span<const pg::VoxelGraphFunction::NodeTypeID>()
+	);
 
 	ERR_FAIL_COND_V_MSG(!result.success, false, result.message);
 
@@ -1535,8 +1681,8 @@ VoxelSingleValue VoxelGeneratorGraph::generate_single(Vector3i position, unsigne
 
 // Note, this wrapper may not be used for main generation tasks.
 // It is mostly used as a debug tool.
-math::Interval VoxelGeneratorGraph::debug_analyze_range(
-		Vector3i min_pos, Vector3i max_pos, bool optimize_execution_map) const {
+math::Interval VoxelGeneratorGraph::debug_analyze_range(Vector3i min_pos, Vector3i max_pos, bool optimize_execution_map)
+		const {
 	std::shared_ptr<const Runtime> runtime_ptr;
 	{
 		RWLockRead rlock(_runtime_lock);
@@ -1546,8 +1692,13 @@ math::Interval VoxelGeneratorGraph::debug_analyze_range(
 	Cache &cache = get_tls_cache();
 	const pg::Runtime &runtime = runtime_ptr->runtime;
 
-	QueryInputs query_inputs(*runtime_ptr, math::Interval(min_pos.x, max_pos.x), math::Interval(min_pos.y, max_pos.y),
-			math::Interval(min_pos.z, max_pos.z), math::Interval());
+	QueryInputs query_inputs(
+			*runtime_ptr,
+			math::Interval(min_pos.x, max_pos.x),
+			math::Interval(min_pos.y, max_pos.y),
+			math::Interval(min_pos.z, max_pos.z),
+			math::Interval()
+	);
 
 	// Note, buffer size is irrelevant here, because range analysis doesn't use buffers
 	runtime.prepare_state(cache.state, 1, false);
@@ -1576,7 +1727,9 @@ math::Interval VoxelGeneratorGraph::debug_analyze_range(
 // Debug land
 
 float VoxelGeneratorGraph::debug_measure_microseconds_per_voxel(
-		bool singular, StdVector<NodeProfilingInfo> *node_profiling_info) {
+		bool singular,
+		StdVector<NodeProfilingInfo> *node_profiling_info
+) {
 	std::shared_ptr<const Runtime> runtime_ptr;
 	{
 		RWLockRead rlock(_runtime_lock);
@@ -1804,68 +1957,83 @@ void VoxelGeneratorGraph::load_graph_from_variant_data(Dictionary data) {
 }
 
 void VoxelGeneratorGraph::_bind_methods() {
-	ClassDB::bind_method(D_METHOD("clear"), &VoxelGeneratorGraph::clear);
-	ClassDB::bind_method(D_METHOD("get_main_function"), &VoxelGeneratorGraph::get_main_function);
+	using Self = VoxelGeneratorGraph;
 
-	ClassDB::bind_method(D_METHOD("set_sdf_clip_threshold", "threshold"), &VoxelGeneratorGraph::set_sdf_clip_threshold);
-	ClassDB::bind_method(D_METHOD("get_sdf_clip_threshold"), &VoxelGeneratorGraph::get_sdf_clip_threshold);
+	ClassDB::bind_method(D_METHOD("clear"), &Self::clear);
+	ClassDB::bind_method(D_METHOD("get_main_function"), &Self::get_main_function);
+
+	ClassDB::bind_method(D_METHOD("set_sdf_clip_threshold", "threshold"), &Self::set_sdf_clip_threshold);
+	ClassDB::bind_method(D_METHOD("get_sdf_clip_threshold"), &Self::get_sdf_clip_threshold);
+
+	ClassDB::bind_method(D_METHOD("is_using_optimized_execution_map"), &Self::is_using_optimized_execution_map);
+	ClassDB::bind_method(D_METHOD("set_use_optimized_execution_map", "use"), &Self::set_use_optimized_execution_map);
+
+	ClassDB::bind_method(D_METHOD("set_use_subdivision", "use"), &Self::set_use_subdivision);
+	ClassDB::bind_method(D_METHOD("is_using_subdivision"), &Self::is_using_subdivision);
+
+	ClassDB::bind_method(D_METHOD("set_subdivision_size", "size"), &Self::set_subdivision_size);
+	ClassDB::bind_method(D_METHOD("get_subdivision_size"), &Self::get_subdivision_size);
+
+	ClassDB::bind_method(D_METHOD("set_debug_clipped_blocks", "enabled"), &Self::set_debug_clipped_blocks);
+	ClassDB::bind_method(D_METHOD("is_debug_clipped_blocks"), &Self::is_debug_clipped_blocks);
+
+	ClassDB::bind_method(D_METHOD("set_use_xz_caching", "enabled"), &Self::set_use_xz_caching);
+	ClassDB::bind_method(D_METHOD("is_using_xz_caching"), &Self::is_using_xz_caching);
+
+	ClassDB::bind_method(D_METHOD("compile"), &Self::_b_compile);
+
+	// ClassDB::bind_method(D_METHOD("generate_single"), &Self::_b_generate_single);
+	ClassDB::bind_method(D_METHOD("debug_analyze_range", "min_pos", "max_pos"), &Self::_b_debug_analyze_range);
 
 	ClassDB::bind_method(
-			D_METHOD("is_using_optimized_execution_map"), &VoxelGeneratorGraph::is_using_optimized_execution_map);
+			D_METHOD("bake_sphere_bumpmap", "im", "ref_radius", "sdf_min", "sdf_max"), &Self::bake_sphere_bumpmap
+	);
 	ClassDB::bind_method(
-			D_METHOD("set_use_optimized_execution_map", "use"), &VoxelGeneratorGraph::set_use_optimized_execution_map);
+			D_METHOD("bake_sphere_normalmap", "im", "ref_radius", "strength"), &Self::bake_sphere_normalmap
+	);
 
-	ClassDB::bind_method(D_METHOD("set_use_subdivision", "use"), &VoxelGeneratorGraph::set_use_subdivision);
-	ClassDB::bind_method(D_METHOD("is_using_subdivision"), &VoxelGeneratorGraph::is_using_subdivision);
-
-	ClassDB::bind_method(D_METHOD("set_subdivision_size", "size"), &VoxelGeneratorGraph::set_subdivision_size);
-	ClassDB::bind_method(D_METHOD("get_subdivision_size"), &VoxelGeneratorGraph::get_subdivision_size);
-
+	ClassDB::bind_method(D_METHOD("debug_load_waves_preset"), &Self::debug_load_waves_preset);
 	ClassDB::bind_method(
-			D_METHOD("set_debug_clipped_blocks", "enabled"), &VoxelGeneratorGraph::set_debug_clipped_blocks);
-	ClassDB::bind_method(D_METHOD("is_debug_clipped_blocks"), &VoxelGeneratorGraph::is_debug_clipped_blocks);
-
-	ClassDB::bind_method(D_METHOD("set_use_xz_caching", "enabled"), &VoxelGeneratorGraph::set_use_xz_caching);
-	ClassDB::bind_method(D_METHOD("is_using_xz_caching"), &VoxelGeneratorGraph::is_using_xz_caching);
-
-	ClassDB::bind_method(D_METHOD("compile"), &VoxelGeneratorGraph::_b_compile);
-
-	// ClassDB::bind_method(D_METHOD("generate_single"), &VoxelGeneratorGraph::_b_generate_single);
-	ClassDB::bind_method(
-			D_METHOD("debug_analyze_range", "min_pos", "max_pos"), &VoxelGeneratorGraph::_b_debug_analyze_range);
-
-	ClassDB::bind_method(D_METHOD("bake_sphere_bumpmap", "im", "ref_radius", "sdf_min", "sdf_max"),
-			&VoxelGeneratorGraph::bake_sphere_bumpmap);
-	ClassDB::bind_method(D_METHOD("bake_sphere_normalmap", "im", "ref_radius", "strength"),
-			&VoxelGeneratorGraph::bake_sphere_normalmap);
-
-	ClassDB::bind_method(D_METHOD("debug_load_waves_preset"), &VoxelGeneratorGraph::debug_load_waves_preset);
-	ClassDB::bind_method(D_METHOD("debug_measure_microseconds_per_voxel", "use_singular_queries"),
-			&VoxelGeneratorGraph::_b_debug_measure_microseconds_per_voxel);
+			D_METHOD("debug_measure_microseconds_per_voxel", "use_singular_queries"),
+			&Self::_b_debug_measure_microseconds_per_voxel
+	);
 
 	// Still present here for compatibility
-	ClassDB::bind_method(D_METHOD("_set_graph_data", "data"), &VoxelGeneratorGraph::load_graph_from_variant_data);
-	ClassDB::bind_method(D_METHOD("_get_graph_data"), &VoxelGeneratorGraph::get_graph_as_variant_data);
+	ClassDB::bind_method(D_METHOD("_set_graph_data", "data"), &Self::load_graph_from_variant_data);
+	ClassDB::bind_method(D_METHOD("_get_graph_data"), &Self::get_graph_as_variant_data);
 
 #ifdef TOOLS_ENABLED
-	ClassDB::bind_method(D_METHOD("_dummy_function"), &VoxelGeneratorGraph::_b_dummy_function);
+	ClassDB::bind_method(D_METHOD("_dummy_function"), &Self::_b_dummy_function);
 #endif
 
-	ADD_PROPERTY(PropertyInfo(Variant::DICTIONARY, "graph_data", PROPERTY_HINT_NONE, "",
-						 PROPERTY_USAGE_NO_EDITOR | PROPERTY_USAGE_INTERNAL),
-			"_set_graph_data", "_get_graph_data");
+	ADD_PROPERTY(
+			PropertyInfo(
+					Variant::DICTIONARY,
+					"graph_data",
+					PROPERTY_HINT_NONE,
+					"",
+					PROPERTY_USAGE_NO_EDITOR | PROPERTY_USAGE_INTERNAL
+			),
+			"_set_graph_data",
+			"_get_graph_data"
+	);
 
 	ADD_GROUP("Performance Tuning", "");
 
 	ADD_PROPERTY(
-			PropertyInfo(Variant::FLOAT, "sdf_clip_threshold"), "set_sdf_clip_threshold", "get_sdf_clip_threshold");
-	ADD_PROPERTY(PropertyInfo(Variant::BOOL, "use_optimized_execution_map"), "set_use_optimized_execution_map",
-			"is_using_optimized_execution_map");
+			PropertyInfo(Variant::FLOAT, "sdf_clip_threshold"), "set_sdf_clip_threshold", "get_sdf_clip_threshold"
+	);
+	ADD_PROPERTY(
+			PropertyInfo(Variant::BOOL, "use_optimized_execution_map"),
+			"set_use_optimized_execution_map",
+			"is_using_optimized_execution_map"
+	);
 	ADD_PROPERTY(PropertyInfo(Variant::BOOL, "use_subdivision"), "set_use_subdivision", "is_using_subdivision");
 	ADD_PROPERTY(PropertyInfo(Variant::INT, "subdivision_size"), "set_subdivision_size", "get_subdivision_size");
 	ADD_PROPERTY(PropertyInfo(Variant::BOOL, "use_xz_caching"), "set_use_xz_caching", "is_using_xz_caching");
 	ADD_PROPERTY(
-			PropertyInfo(Variant::BOOL, "debug_block_clipping"), "set_debug_clipped_blocks", "is_debug_clipped_blocks");
+			PropertyInfo(Variant::BOOL, "debug_block_clipping"), "set_debug_clipped_blocks", "is_debug_clipped_blocks"
+	);
 
 	ADD_SIGNAL(MethodInfo(SIGNAL_NODE_NAME_CHANGED, PropertyInfo(Variant::INT, "node_id")));
 }